requirements.dev.txt
flake8>=7.0.0,<7.1
pytest>=8.2.0,<8.3
pytest-django>=4.8.0,<4.9
pytest-cov>=5.0.0,<5.1
django-silk==5.2.0